Course rigor improvements and quiz fix

+++ courses/database-implementation/03-memory-pools-allocators/solution/cpp.cpp
@@ -95,9 +95,12 @@ private:
 // ---------------------------------------------------------------------------
 
 int main() {
-    SlabAllocator<int, 8> alloc;
+    // T must be at least pointer-sized so a free slot can hold the
+    // next-pointer of the free list (see the static_assert above), so the
+    // demo uses int64_t rather than int.
+    SlabAllocator<int64_t, 8> alloc;
 
-    std::array<int*, 4> ptrs{};
+    std::array<int64_t*, 4> ptrs{};
     for (int i = 0; i < 4; ++i) {
         ptrs[i] = alloc.allocate();
         *ptrs[i] = i * 10;
@@ -106,8 +109,8 @@ int main() {
     alloc.deallocate(ptrs[1]);
     alloc.deallocate(ptrs[2]);
 
-    int* r0 = alloc.allocate();
-    int* r1 = alloc.allocate();
+    int64_t* r0 = alloc.allocate();
+    int64_t* r1 = alloc.allocate();
 
     bool reused = (r0 == ptrs[2] || r0 == ptrs[1]) &&
                   (r1 == ptrs[2] || r1 == ptrs[1]);
